feat: include source files and declaration maps
This is a much cleaner approach than source mapping and
works well with lerna.

They are excluded in the production build because seeing the source code
isn't very helpful for users of the package. If they find a bug it will be easier
to debug in their own clone of this repository